remove new invite modal ff
