+++ static/app/components/modals/inviteMembersModal/index.tsx
@@ -11,15 +11,13 @@ import {
 } from 'sentry/components/modals/inviteMembersModal/inviteHeaderMessages';
 import {InviteMembersContext} from 'sentry/components/modals/inviteMembersModal/inviteMembersContext';
 import InviteMembersFooter from 'sentry/components/modals/inviteMembersModal/inviteMembersFooter';
-import InviteMembersModalView from 'sentry/components/modals/inviteMembersModal/inviteMembersModalview';
-import InviteRowControl from 'sentry/components/modals/inviteMembersModal/inviteRowControlNew';
+import InviteRowControl from 'sentry/components/modals/inviteMembersModal/inviteRowControl';
 import type {InviteRow} from 'sentry/components/modals/inviteMembersModal/types';
 import useInviteModal from 'sentry/components/modals/inviteMembersModal/useInviteModal';
 import {InviteModalHook} from 'sentry/components/modals/memberInviteModalCustomization';
 import {ORG_ROLES} from 'sentry/constants';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {trackAnalytics} from 'sentry/utils/analytics';
 import {isActiveSuperuser} from 'sentry/utils/isActiveSuperuser';
 import useOrganization from 'sentry/utils/useOrganization';
 
@@ -31,7 +29,6 @@ interface InviteMembersModalProps extends ModalRenderProps {
 function InviteMembersModal({
   Header,
   Body,
-  closeModal,
   initialData,
   source,
   Footer,
@@ -39,13 +36,10 @@ function InviteMembersModal({
   const organization = useOrganization();
 
   const {
-    addInviteRow,
     invites,
     memberResult,
-    removeInviteRow,
     reset,
     sendInvites,
-    sessionId,
     setEmails,
     setRole,
     setTeams,
@@ -88,7 +82,7 @@ function InviteMembersModal({
           headerInfo: headerInfo,
           isOverMemberLimit: isOverMemberLimit,
         }) => {
-          return organization.features.includes('invite-members-new-modal') ? (
+          return (
             <InviteMembersContext.Provider
               value={{
                 willInvite,
@@ -120,40 +114,9 @@ function InviteMembersModal({
                 />
               </Body>
               <Footer>
-                <InviteMembersFooter canSend />
+                <InviteMembersFooter canSend={canSend} />
               </Footer>
             </InviteMembersContext.Provider>
-          ) : (
-            <InviteMembersModalView
-              addInviteRow={addInviteRow}
-              canSend={canSend}
-              closeModal={() => {
-                trackAnalytics('invite_modal.closed', {
-                  organization,
-                  modal_session: sessionId,
-                });
-                closeModal();
-              }}
-              complete={complete}
-              Footer={Footer}
-              headerInfo={headerInfo}
-              invites={invites}
-              inviteStatus={inviteStatus}
-              isOverMemberLimit={
-                isOverMemberLimit && organization.features.includes('invite-billing')
-              }
-              member={memberResult.data}
-              pendingInvites={pendingInvites}
-              removeInviteRow={removeInviteRow}
-              reset={reset}
-              sendingInvites={sendingInvites}
-              sendInvites={inviteModalSendInvites}
-              setEmails={setEmails}
-              setRole={setRole}
-              setTeams={setTeams}
-              willInvite={willInvite}
-              error={error}
-            />
           );
         }}
       </InviteModalHook>
