+++ static/app/components/modals/inviteMembersModal/inviteRowControl.spec.tsx
@@ -7,7 +7,7 @@ import {
   InviteMembersContext,
   type InviteMembersContextValue,
 } from 'sentry/components/modals/inviteMembersModal/inviteMembersContext';
-import InviteRowControlNew from 'sentry/components/modals/inviteMembersModal/inviteRowControlNew';
+import InviteRowControlNew from 'sentry/components/modals/inviteMembersModal/inviteRowControl';
 import TeamStore from 'sentry/stores/teamStore';
 import type {DetailedTeam} from 'sentry/types/organization';
