+++ static/app/utils/analytics/growthAnalyticsEvents.tsx
@@ -119,11 +119,7 @@ export type GrowthEventParameters = {
   'growth.submitted_mobile_prompt_ask_teammate': MobilePromptBannerParams;
   'invite_modal.add_more': InviteModal;
   'invite_modal.closed': InviteModal;
-  'invite_modal.invites_sent': InviteModal & {
-    failed_invites: number;
-    is_new_modal: boolean;
-    sent_invites: number;
-  };
+  'invite_modal.invites_sent': InviteModal;
   'invite_modal.opened': InviteModal & {
     can_invite: boolean;
     source?: string;
